+++ src/presentation/helpers/http-helper.ts
@@ -1,5 +1,5 @@
 import { ServerError } from '../errors'
-import { type IHttpResponse } from '../protocols/http'
+import { type IHttpResponse } from '../protocols'
 
 export function badRequest (error: Error): IHttpResponse {
   return {
